app/api_service/requirements.txt
asgiref==3.3.1
boto3==1.17.6
botocore==1.20.6
Django==3.1.6
djangorestframework==3.12.2
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
